lab8/airflow-example1/requirements.txt
pandas
apache-airflow==2.4.3
apache-airflow-providers-google==8.4.0